pyproject.toml
[tool.poetry]
name = "devman-django"
version = "0.1.0"
description = ""
authors = ["Aleksandr Karakchiev <realrushen@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8.0"
Django = "^3.2.3"
Pillow = "^8.2.0"
django-admin-sortable2 = "^1.0"
django-tinymce = "^3.3.0"
environs = "^9.3.2"
dj-database-url = "^0.5.0"
requests = "^2.25.1"
pydantic = "^1.8.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
